requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
Pillow
psycopg2-binary
gunicorn
whitenoise
requests
